requirements.txt
pg8000
configparser
numpy
pandas
sentence_transformers
tqdm
faiss-cpu
transformers
streamlit